Enable sound output from SVGAlib-Vectrex games
Sound support is enabled but games are designed for the old OSS sound
system. To emulate this with alsa, install the alsa-oss package and
start with the aoss command, eg "sudo aoss zblast -nomusic"

The interrupt disabling messes up the sound a bit, so -nomusic is
recommended so that it doesn't sound too bad in Zblast.