Use baremetal menu in premade image; copy files from release
This enables the baremetal piTrexBoot menu on initial boot on
the packer build OS image.

I've tested it with the menu starting Raspbian & helloWorld.

I've not tried anything wrt Vecx as yet, but will probably move onto
that once this is merged.

To prevent the need to build everything to get a working OS image, i've
opted to use the 'pre-cut' tgz from the github releases page.

I create a compiled version of the menu system though, as this is the
area that needs the most iteration - and so worth building from master.