Fix interrupt disabling code for Raspberry Pi Zero 2 W.
Display should be completely clean on the Pi Zero 2 running Linux
if the CPU governer is set to "performance", "sudo tvservice -o" is
run, and the kernel is booted with the "nosmp" boot code to disable
running multiple CPU cores.

Infrequent display glitches occour when multiple CPU cores are
enabled.