Add scripts and other files for PiTrexCore Linux environment.
Includes files for the Raspberry Pi system, and, in the distro/PC
directory, the x86 system for running XMAME remotely. See README
files for details on what the individual scripts are for.

distro/tcz/premade contains extensions (currently one) which are
not made automatically by the distro/piext.sh script.